Limpiando notificade, Peers y eliminando
