+++ js/main.js
@@ -1,29 +1,25 @@
-import LoginUser from "./login.js";
 import RegisterUser from "./register.js";
 import User from './User.js'
 import Authenticated from "./authenticated.js";
-import Notificades from "./notificade.js";
-import Feedback from "./Feedback.js";
+import Skills from "./evaluate-skills.js";
 import UI from './UI.js';
-// import newPassword from './newPassword.js';
-import Invitated from './invitated.js';
-import Peers from './peers.js';
+import Feedback from './feedback.js'
 
 
 const user = new User;
+const feedback = new Feedback;
+const authenticated = new Authenticated();
 
-const login = new LoginUser();
-login.authenticated();
 
-const authenticatedIndex = new Authenticated();
-authenticatedIndex.verificarLogin();
+authenticated.authenticated();
+authenticated.verificarLogin();
 
 
 if (window.location.pathname === "/views/login.html") {
 	document.getElementById("login").addEventListener("submit", (event) => {
 		event.preventDefault();
 
-        login.loginUsers()
+        authenticated.loginUsers()
     })
 }
 
@@ -41,7 +37,7 @@ if (window.location.pathname === "/index.html" || window.location.pathname === '
 		user.mostrarMenuCambiarContraseña(); //controlador de vistas
 
 		document.getElementById('logout').addEventListener('click', (go) =>{ //evento para ancla
-			authenticatedIndex.finalizarSesion() //Re-utilizamos la clase agregando nuevos metos
+			authenticated.finalizarSesion() //Re-utilizamos la clase agregando nuevos metos
 			window.location.reload(); //recargamos para finalizar sesion
 		})
 		document.getElementById('form-change-password').addEventListener('submit', (event) =>{ 
@@ -55,8 +51,7 @@ if (window.location.pathname === "/index.html" || window.location.pathname === '
 		document.getElementById('form-invite-feedback').addEventListener('submit', (event) =>{ 
 			//evento para ancla
 			event.preventDefault();
-			const invited = new Invitated;
-			invited.inviteUsers();
+			feedback.inviteUsers();
 		})
 }
 
@@ -72,24 +67,13 @@ if (window.location.pathname === "/index.html" || window.location.pathname === '
 //notificade
 document.getElementById("note").addEventListener("click",(event)=>{
 	event.preventDefault();
-	const notificade = new Notificades();
-	notificade.bodyDiv();
+	feedback.bodyDiv();
 });
-
-
-// if(window.location.pathname === '/index.html') {
-//     const invitated = new Invitated;
-//     invitated.inviteUsers();
-// }
-
-// const register = new RegisterUser
-
-// const register = new RegisterUser
 
 //************************** PEERS *************************************/
 document.getElementById("peers").addEventListener("click",(event)=>{
 	event.preventDefault();
-	const peers = new Peers();
+	// const peers = new Peers();
 	//peers.Entro();
-	peers.first()
+	feedback.first()
 });
